chore(cspc-operator) : containerise cspc-operator to run as a deployment.
This commit decouples cspc operator from maya-apiserver.
